pyproject.toml
[tool.poetry]
name = "orgs"
version = "0.1.0"
description = ""
authors = ["Vitor Toledo <vitor.toledo@grupomateus.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
makefile = "^1.1.0"
django-easy-tenants = "^0.9.2"
ruff = "^0.8.4"
pre-commit = "^4.0.1"
django-cpf-cnpj = "^1.0.0"
django-simple-history = "^3.7.0"
xlsxwriter = "^3.2.0"
django-health-check = "^3.18.3"
fpdf2 = "^2.8.2"
ofxtools = "^0.9.5"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"
google-cloud-secret-manager = "^2.23.0"
google-cloud-storage = "^3.0.0"
django-storages = {extras = ["google"], version = "^1.14.5"}
sendgrid = "^6.11.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
ipython = "^9.0.2"
phonenumbers = "^8.13.30"
django-redis = "^5.4.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
